+++ lib/ui/Gazetteer.mjs
@@ -1,18 +1,4 @@
-export default params => {
-
-  // Assign params to gazetteer object.
-  const gazetteer = Object.assign({
-
-    getLocation,
-
-    sources: {
-      glx: glx,
-      mapbox: mapbox,
-      google: google
-    }
-
-  }, params);
-
+export default gazetteer => {
 
   gazetteer.input = mapp.utils.html.node`<input 
     type="search" 
@@ -23,7 +9,6 @@ export default params => {
   gazetteer.node = mapp.utils.html.node`<div class="dropdown">
     ${gazetteer.input}
     ${gazetteer.list}`
-
 
   gazetteer.target.append(gazetteer.node)
 
@@ -36,24 +21,19 @@ export default params => {
     gazetteer.list.innerHTML = ''
 
     // Only search if value has length.
-    e.target.value.length > 0 && search(e.target.value)
-  })
+    if (!e.target.value.length) return;
 
-  function search(term){
- 
     // Get possible coordinates from input.
-    let ll = term.split(',').map(parseFloat)
+    let ll = e.target.value.split(',').map(parseFloat)
 
     // Check whether coordinates are float values.
     if (ll.length === 2 && ll.every(n => !isNaN(n))) {
 
       gazetteer.list.appendChild(mapp.utils.html.node`
-      <li 
-        onclick=${e=>{
+          <li 
+            onclick=${e => {
 
-          gazetteer.list.innerHTML = ''
-
-          gazetteer.getLocation({
+          mapp.utils.gazetteer.getLocation({
             label: `Latitude:${ll[0]}, Longitude:${ll[1]}`,
             source: 'Coordinates',
             lng: ll[1],
@@ -66,42 +46,18 @@ export default params => {
       return;
     }
 
-    mapp.utils.AutocompleteService.getPlacePredictions({
-      input: term
-    }).then(response => {
+    // Validate whether gazetteer service is available.
+    if (!Object.hasOwn(mapp.utils.gazetteer, gazetteer.provider)) {
 
-      response.predictions.forEach(prediction => {
+      return console.warn('Requested gazetteer service not available');
+    }
 
-        gazetteer.list.append(mapp.utils.html.node`
-        <li
-          onclick=${e=>{
+    mapp.utils.gazetteer[gazetteer.provider](e.target.value, gazetteer)
 
-            //if (gazetteer.callback) return gazetteer.callback(entry)
+  })
 
-            mapp.utils.PlacesService.getDetails({
-              placeId: prediction.place_id,
-              fields: ['geometry']
-            }, (place, status) => {
-              console.log(place)
-              gazetteer.getLocation({
-                label: 'Foo',
-                source: 'Google',
-                lng: place.geometry.location.lng(),
-                lat: place.geometry.location.lat()
-              })
-            });
-    
-          }}>
-          <span class="label">Google</span>
-          <span>${prediction.description}</span>`)
-
-
-
-      })
-    })
-
-    return;
-
+  function search(term){
+ 
     // Abort current xhr and create new.
     gazetteer.xhr = new XMLHttpRequest()
   
@@ -160,90 +116,5 @@ export default params => {
   
     gazetteer.xhr.send()
   }
-
-  function glx(record) {
-
-    gazetteer.getLocation({
-      layer: gazetteer.mapview.layers[record.layer],
-      table: record.table,
-      id: record.id
-    }, loc => mapp.location.get(loc).then(loc => loc && loc.flyTo())) 
-  }
   
-  function google(record) {
-
-    mapp.utils.PlacesService.getDetails({
-      placeId: record.id,
-      fields: ['geometry']
-    }, (place, status) => {
-      console.log(place)
-      gazetteer.getLocation({
-        label: record.label,
-        source: record.source,
-        lng: place.geometry.location.lng(),
-        lat: place.geometry.location.lat()
-      })
-    });
-
-  }
-
-  function mapbox(record) {
-
-    gazetteer.getLocation({
-      label: record.label,
-      source: record.source,
-      lng: record.marker[0],
-      lat: record.marker[1]
-    })
-  }
-
-  function getLocation(location, callback) {
-
-    if (typeof callback === 'function') {
-      callback(location)
-      return;
-    }
-
-    Object.assign(location, {
-      layer: {
-        mapview: gazetteer.mapview
-      },
-      Layers: [],
-      hook: location.label
-    })
-
-    const infoj = [
-      {
-        title: location.label,
-        value: location.source,
-        inline: true
-      },
-      {
-        type: 'pin',
-        value: [location.lng, location.lat],
-        srid: '4326',
-        class: 'display-none',
-        location
-      }
-    ]
-
-    gazetteer.streetview && infoj.push({ 
-      type: 'streetview',
-      location
-    })
-
-    mapp.location.decorate(Object.assign(location, { infoj }))
-
-    gazetteer.mapview.locations[location.hook] = location
-
-    if (gazetteer.zoom) {
-      let view = gazetteer.mapview.Map.getView()
-      view.setZoom(gazetteer.zoom)
-      view.setCenter(ol.proj.fromLonLat([location.lng, location.lat]))
-      
-    } else {
-      location.flyTo()
-    }
-  }
-
 }
