+++ lib/utils/_utils.mjs
@@ -54,6 +54,8 @@ import style from './style.mjs'
 
 import * as svgSymbols from './svgSymbols.mjs'
 
+import * as gazetteer from './gazetteer.mjs'
+
 import {default as verticeGeoms} from './verticeGeoms.mjs'
 
 import xhr from './xhr.mjs'
@@ -69,6 +71,7 @@ export default {
   copyToClipboard,
   csvDownload,
   dataURLtoBlob,
+  gazetteer,
   getCurrentPosition,
   hexa,
   loadPlugins,
